refactor(leveling): unsigned XP types and discriminated notification targets
- Switch all XP/level arithmetic from i64 to u64/u32 (non-negative
  invariant enforced at the type level).
- Replace NotificationScope enum with a discriminated union
  NotificationTarget where SpecifiedChannel carries channelId,
  eliminating the previous Option + scope mismatch.
- LevelingScope, LevelReward, and XpMultiplier redesigned with
  a `target` discriminated union for role vs channel targets.
- Queries now return strongly-typed results instead of raw i64,
  with cast helpers for sqlx boundaries.

BREAKING CHANGE: database query signatures changed; the
corresponding dashboard types must be updated to match.

+++ src/features/leveling/placeholders.rs
@@ -1,11 +1,11 @@
 use crate::core::config::guild_ctx::GuildCtx;
-use crate::shared::placeholders::{DiscordCtx, PlaceholderResolver, ResolverChain, render};
+use crate::shared::placeholders::{render, DiscordCtx, PlaceholderResolver, ResolverChain};
 use serenity::all::User;
 
 /// Custom resolver for leveling-specific keys like `{level}` or `{level.old}`
 pub struct LevelingResolver {
-    pub current_level: i64,
-    pub previous_level: i64,
+    pub current_level: u32,
+    pub previous_level: u32,
 }
 
 impl PlaceholderResolver for LevelingResolver {
@@ -22,8 +22,8 @@ pub fn replace_level_notify_placeholder(
     text: &str,
     gctx: &GuildCtx,
     user: &User,
-    current_level: i64,
-    previous_level: i64,
+    current_level: u32,
+    previous_level: u32,
 ) -> String {
     let discord_ctx = DiscordCtx {
         gctx: Some(gctx),
